Add Twitch emote responses
This pulls a list of emotes Twitchemotes.com API and will respond to
messages matching an emote name by downloading the emote from Twitch
(cached to only download each emote once per session) and sending it to
Discord. This way the response is just the emote isolated by itself.

It's not a great solution, but given the current constraints with the
Discord API, I think this is the best approach to it.

Closes #4